refactor/ JPQL 을 통한 최적화

+++ src/test/kotlin/com/group/libraryapp/service/book/BookServiceTest.kt
@@ -152,6 +152,6 @@ class BookServiceTest @Autowired constructor(
     }
 }
 
-private fun assertCount(results: List<BookStatResponse>, type: BookType, count: Int) {
+private fun assertCount(results: List<BookStatResponse>, type: BookType, count: Long) {
     assertThat(results.first { it.type == type}.count).isEqualTo(count)
 }
